Fix pkg example express app (#668)
* Fix pkg example express app

Currently, following the instructions provided in the README result in pkg failing to build the executable. The error output references "node7" as being an undefined target in package.json.

pkg relies on pkg-fetch for retrieving Node versions to be used as targets in building executables. As of pkg-fetch v2.5, no node v7.x.x is listed as an available source. Changing "node7" to "node10" fixes this issue as it is available via pkg-fetch v2.5.

Additionally, this PR:
*) Adds files/folders to .gitignore related to the example express app which should not be included in master.
*) Improves README ducumentation for example express app.

* Revise example app node target version

Changed `pkg` target from `node10` to `node8` in order to pass CI build errors due to missing peer dependency for babel at that version.