Use "linuxstatic" to generate bytecode for different arch on Linux
Currently `pkg` can't generate executable for different arch:

> Warning Failed to make bytecode node14-arm64 for file ...

as it attempts to run binary of different arch.

We can set up QEMU and binfmt to allow emulation. However, the "linux"
binary needs libraries from the system, and it is PITA to set up the
entire sysroot of a different arch.

This change workarounds the issue by using "linuxstatic" to generate
bytecode for different arch.